+++ packages/perseus-editor/src/components/__tests__/locked-point-settings.test.tsx
@@ -8,7 +8,11 @@ import {getDefaultFigureForType} from "../util";
 
 import type {UserEvent} from "@testing-library/user-event";
 
-const defaultProps = getDefaultFigureForType("point");
+const defaultProps = {
+    ...getDefaultFigureForType("point"),
+    onRemove: () => {},
+    onChangeProps: () => {},
+};
 
 describe("LockedPointSettings", () => {
     let userEvent: UserEvent;
@@ -27,138 +31,11 @@ describe("LockedPointSettings", () => {
         );
 
         const titleText = screen.getByText("Point (0, 0)");
-        const colorSwatch = screen.getByLabelText("Color: grayH, filled");
+        const colorSwatch = screen.getByLabelText("grayH, filled");
 
         // Assert
         expect(titleText).toBeInTheDocument();
         expect(colorSwatch).toBeInTheDocument();
-    });
-
-    test("Should not show the color in summary if toggled off", () => {
-        // Arrange
-
-        // Act
-        render(
-            <LockedPointSettings
-                {...defaultProps}
-                onChangeProps={() => {}}
-                toggled={false}
-                onToggle={() => {}}
-            />,
-            {wrapper: RenderStateRoot},
-        );
-
-        const titleText = screen.getByText("Point (0, 0)");
-        const colorSwatch = screen.queryByLabelText(
-            "Point color: blue, filled",
-        );
-
-        // Assert
-        expect(titleText).toBeInTheDocument();
-        expect(colorSwatch).not.toBeInTheDocument();
-    });
-
-    test("Should show toggle switch if onToggle is passed in", () => {
-        // Arrange
-
-        // Act
-        render(
-            <LockedPointSettings
-                {...defaultProps}
-                onChangeProps={() => {}}
-                onToggle={() => {}}
-            />,
-            {wrapper: RenderStateRoot},
-        );
-
-        const toggleSwitch = screen.getByLabelText("show point on graph");
-
-        // Assert
-        expect(toggleSwitch).toBeInTheDocument();
-    });
-
-    test("Toggle switch should match toggled prop when true", () => {
-        // Arrange
-
-        // Act
-        render(
-            <LockedPointSettings
-                {...defaultProps}
-                onChangeProps={() => {}}
-                toggled={true}
-                onToggle={() => {}}
-            />,
-            {wrapper: RenderStateRoot},
-        );
-
-        const toggleSwitch = screen.getByLabelText("show point on graph");
-
-        // Assert
-        expect(toggleSwitch).toBeChecked();
-    });
-
-    test("Toggle switch should match toggled prop when false", () => {
-        // Arrange
-
-        // Act
-        render(
-            <LockedPointSettings
-                {...defaultProps}
-                onChangeProps={() => {}}
-                toggled={false}
-                onToggle={() => {}}
-            />,
-            {wrapper: RenderStateRoot},
-        );
-
-        const toggleSwitch = screen.getByLabelText("show point on graph");
-
-        // Assert
-        expect(toggleSwitch).not.toBeChecked();
-    });
-
-    test("Should show extra fields if toggled on", () => {
-        // Arrange
-
-        // Act
-        render(
-            <LockedPointSettings
-                {...defaultProps}
-                onChangeProps={() => {}}
-                toggled={true}
-                onToggle={() => {}}
-            />,
-            {wrapper: RenderStateRoot},
-        );
-
-        const colorSelect = screen.getByLabelText("color");
-        const openCheckbox = screen.getByLabelText("open point");
-
-        // Assert
-        expect(colorSelect).toBeInTheDocument();
-        expect(openCheckbox).toBeInTheDocument();
-    });
-
-    test("Should not show extra fields if toggled off", () => {
-        // Arrange
-
-        // Act
-        render(
-            <LockedPointSettings
-                {...defaultProps}
-                onChangeProps={() => {}}
-                toggled={false}
-                onToggle={() => {}}
-            />,
-            {wrapper: RenderStateRoot},
-        );
-
-        const colorSelect = screen.queryByLabelText("Color");
-        const openCheckbox = screen.queryByLabelText("Open point");
-
-        // Assert
-        expect(colorSelect).not.toBeInTheDocument();
-        expect(openCheckbox).not.toBeInTheDocument();
     });
 
     test("Summary should reflect the coordinates", () => {
@@ -253,5 +130,24 @@ describe("LockedPointSettings", () => {
                 expect(coordField).toHaveValue(expectedValue);
             },
         );
+    });
+
+    test("Calls onToggle when header is clicked", async () => {
+        // Arrange
+        const onToggle = jest.fn();
+        render(
+            <LockedPointSettings
+                {...defaultProps}
+                onChangeProps={() => {}}
+                onToggle={onToggle}
+            />,
+            {wrapper: RenderStateRoot},
+        );
+
+        // Act
+        await userEvent.click(screen.getByText("Point (0, 0)"));
+
+        // Assert
+        expect(onToggle).toHaveBeenCalled();
     });
 });