+++ packages/perseus-editor/src/components/locked-line-settings.tsx
@@ -7,17 +7,18 @@
 import {View, useUniqueIdWithMock} from "@khanacademy/wonder-blocks-core";
 import {OptionItem, SingleSelect} from "@khanacademy/wonder-blocks-dropdown";
 import {Strut} from "@khanacademy/wonder-blocks-layout";
-import {color as wbColor, spacing} from "@khanacademy/wonder-blocks-tokens";
+import {spacing} from "@khanacademy/wonder-blocks-tokens";
 import {LabelMedium, LabelLarge} from "@khanacademy/wonder-blocks-typography";
 import {StyleSheet} from "aphrodite";
 import * as React from "react";
 
 import ColorSelect from "./color-select";
+import DefiningPointSettings from "./defining-point-settings";
 import LineSwatch from "./line-swatch";
 import LockedFigureSettingsAccordion from "./locked-figure-settings-accordion";
 import LockedFigureSettingsActions from "./locked-figure-settings-actions";
-import LockedPointSettings from "./locked-point-settings";
 
+import type {AccordionProps} from "./locked-figure-settings";
 import type {
     LockedFigure,
     LockedFigureColor,
@@ -25,10 +26,17 @@ import type {
     LockedPointType,
 } from "@khanacademy/perseus";
 
-export type Props = LockedLineType & {
-    onRemove: () => void;
-    onChangeProps: (newProps: Partial<LockedFigure>) => void;
-};
+export type Props = LockedLineType &
+    AccordionProps & {
+        /**
+         * Called when the delete button is pressed.
+         */
+        onRemove: () => void;
+        /**
+         * Called when the props (points, color, etc.) are updated.
+         */
+        onChangeProps: (newProps: Partial<LockedFigure>) => void;
+    };
 
 const LockedLineSettings = (props: Props) => {
     const {
@@ -47,7 +55,6 @@ const LockedLineSettings = (props: Props) => {
     // with their respective text fields.
     const ids = useUniqueIdWithMock();
     const kindSelectId = ids.get("line-kind-select");
-    const colorSelectId = ids.get("line-color-select");
     const styleSelectId = ids.get("line-style-select");
 
     const capitalizeKind = kind.charAt(0).toUpperCase() + kind.slice(1);
@@ -87,6 +94,8 @@ const LockedLineSettings = (props: Props) => {
 
     return (
         <LockedFigureSettingsAccordion
+            expanded={props.expanded}
+            onToggle={props.onToggle}
             header={
                 <View style={styles.row}>
                     <LabelLarge>{lineLabel}</LabelLarge>
@@ -114,15 +123,14 @@ const LockedLineSettings = (props: Props) => {
                     placeholder=""
                 >
                     <OptionItem value="line" label="line" />
-                    <OptionItem value="segment" label="segment" />
                     <OptionItem value="ray" label="ray" />
+                    <OptionItem value="segment" label="segment" />
                 </SingleSelect>
             </View>
 
             <View style={[styles.row, styles.spaceUnder]}>
                 {/* Line color settings */}
                 <ColorSelect
-                    id={colorSelectId}
                     selectedValue={lineColor}
                     onChange={handleColorChange}
                 />
@@ -154,21 +162,23 @@ const LockedLineSettings = (props: Props) => {
             </View>
 
             {/* Defining points settings */}
-            <LockedPointSettings
+            <DefiningPointSettings
                 label="Point 1"
-                toggled={showPoint1}
+                showPoint={showPoint1}
                 {...point1}
-                onToggle={(newValue) => onChangeProps({showPoint1: newValue})}
+                onTogglePoint={(newValue) =>
+                    onChangeProps({showPoint1: newValue})
+                }
                 onChangeProps={(newProps) => handleChangePoint(newProps, 0)}
-                style={styles.lockedPointSettingsContainer}
             />
-            <LockedPointSettings
+            <DefiningPointSettings
                 label="Point 2"
-                toggled={showPoint2}
+                showPoint={showPoint2}
                 {...point2}
-                onToggle={(newValue) => onChangeProps({showPoint2: newValue})}
+                onTogglePoint={(newValue) =>
+                    onChangeProps({showPoint2: newValue})
+                }
                 onChangeProps={(newProps) => handleChangePoint(newProps, 1)}
-                style={styles.lockedPointSettingsContainer}
             />
 
             {/* Actions */}
@@ -183,13 +193,6 @@ const LockedLineSettings = (props: Props) => {
 };
 
 const styles = StyleSheet.create({
-    lockedPointSettingsContainer: {
-        marginTop: spacing.xSmall_8,
-        marginBottom: 0,
-        marginLeft: -spacing.xxxSmall_4,
-        marginRight: -spacing.xxxSmall_4,
-        backgroundColor: wbColor.white,
-    },
     row: {
         flexDirection: "row",
         alignItems: "center",